Refactor packs
Basic packs (non-engine packs) no longer create an Engine. They are now
simple additions to the main app's paths.

Engine packs now create the "Engine" class underneath the namespace.
This makes for less surprises.

The Pack object is now just a shell and has a reference to the Engine
(if it's a Engine Pack).