Bugfix: default path does not have the same format as to_run
Using rspec 3.11, I noticed that specs in my packs weren't running. Perhaps something may have changed in Rspec since this integration was committed.

By default, the value of `default_path` is `spec` and the value of `to_run` (when just running the `rspec` command) is `['/full/path/to/spec']`.

After joining `default_path` to `Rails.root` everything seems to be working as expected.